refactor: design & cleanup
