+++ apps/dashboard/package.json
@@ -26,7 +26,7 @@
     "@vercel/speed-insights": "^1.0.10",
     "clsx": "^2.1.1",
     "fuse.js": "^7.0.0",
-    "next": "14.2.3",
+    "next": "14.2.7",
     "next-themes": "^0.3.0",
     "react": "^18",
     "react-colorful": "^5.6.1",
