+++ apps/dashboard/src/app/api/fonts/route.tsx
@@ -1,7 +1,6 @@
 import { getFontData } from "../../../lib/fonts";
 
 export const dynamic = "force-static";
-export const runtime = "nodejs";
 
 export async function GET() {
   return Response.json(await getFontData());
