+++ apps/dashboard/src/stores/fontsStore.ts
@@ -3,17 +3,19 @@ import { DEFAULT_FONTS, type Font } from "../lib/fonts";
 
 interface FontsState {
   installedFonts: Set<string>;
-  /** install the font based on its name */
   installFont: (font: string) => void;
   allFonts: Font[];
 }
 
+// If we are in a browser, fetch the available fonts from the API
 if (typeof window !== "undefined") {
   void fetch("/api/fonts")
-    .then((res) => res.json())
+    .then((response) => response.json())
     .then((fonts) => {
       useFontsStore.getState().allFonts = fonts as Font[];
-    });
+    })
+    // eslint-disable-next-line @typescript-eslint/use-unknown-in-catch-callback-variable, no-console -- the error is logged
+    .catch(console.error);
 }
 
 export const useFontsStore = create<FontsState>((set) => ({
@@ -22,7 +24,6 @@ export const useFontsStore = create<FontsState>((set) => ({
   installFont: (name) => {
     set((state) => {
       state.installedFonts.add(name);
-
       return state;
     });
   },
